feat(): Dependencies for Angular Universal
Added WS and xmlhttprequest as dependencies, as they are required to allow normal functions of the Firebase JS SDK in Universal. Created a new `angularfire/firebase-node` module, which includes these + will include any additional fixes for Universal in the future.

The rest of the SSR work is taking place in #1454.